+++ lib/packages/core/src/typings/assetBalance.ts
@@ -1,0 +1,10 @@
+/** @module core */
+
+/**
+ * Original work Copyright (c) 2019 Burst Apps Team
+ */
+
+export interface AssetBalance {
+    balanceQNT: string;
+    asset: string;
+}