+++ web/angular-wallet/src/app/main/blocks/blocks.component.ts
@@ -62,7 +62,7 @@ export class BlocksComponent {
     }
 
     public isOwnAccount(address: string): boolean {
-        return address != undefined && address == this.account.address;
+        return address != undefined && address == this.account.accountRS;
     }
 
     public convertTimestamp(timestamp: number): Date {