+++ web/angular-wallet/src/app/main/transactions/transactions.component.ts
@@ -67,9 +67,9 @@ export class TransactionsComponent {
     this.dataSource.filter = filterValue || 'burst';
   }
 
-  public isOwnAccount(address: string): boolean {
-    return address !== undefined && address === this.account.address;
-  }
+    public isOwnAccount(address: string): boolean {
+        return address != undefined && address == this.account.accountRS;
+    }
 
   public convertTimestamp(timestamp: number): Date {
     return convertBurstTimeToDate(timestamp);