+++ lib/packages/core/CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file.
     - Block: `getBlocks`
     
  
+
+## Unreleased
+### Account
+- Modified `Account` type to better reflect what is returned from BRS API. This is a breaking change, please see the `Account` for the new property names.  
+
 ## 0.1.0-rc1
 ### Network
 - Added `suggestFees` for getting suggested fees.